# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agentic Brain API",
    "description": "Multi-LLM orchestration platform with GraphRAG, Unified Brain, and real-time chat",
    "version": "1.0.0"
  },
  "paths": {
...
    "/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Send Chat Message",
        "description": "Send a message to the chatbot and receive a response",
        "operationId": "chat_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyQuery": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/chat/stream": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Stream Chat Response",
        "description": "Stream chat response using Server-Sent Events (SSE)",
        "operationId": "stream_chat_chat_stream_get",
        "security": [
          {
            "APIKeyHeader": []
          },
          {
            "APIKeyQuery": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 10000,
              "description": "User message",
              "title": "Message"
            },
            "description": "User message"
          },
          {
            "name": "session_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Session ID",
              "title": "Session Id"
            },
            "description": "Session ID"
          },
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "User ID",
              "title": "User Id"
            },
            "description": "User ID"
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "LLM provider: ollama, openai, or anthropic",
              "default": "ollama",
              "title": "Provider"
            },
            "description": "LLM provider: ollama, openai, or anthropic"
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Model name",
              "default": "llama3.1:8b",
              "title": "Model"
            },
            "description": "Model name"
          },
          {
            "name": "temperature",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 2.0,
              "minimum": 0.0,
              "description": "Sampling temperature",
              "default": 0.7,
              "title": "Temperature"
            },
            "description": "Sampling temperature"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 32000,
            "minLength": 1,
            "title": "Message",
            "description": "The user's message"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "Session ID for conversation tracking"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id",
            "description": "User ID for multi-user support"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata",
            "description": "Optional metadata for extended request information"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "description": "Request model for chat endpoint.\n\nAttributes:\n    message: The user's message/query\n    session_id: Optional session identifier for conversation tracking\n    user_id: Optional user identifier for multi-user support\n    metadata: Optional metadata dictionary for extended request information"
      },
      "ChatResponse": {
        "properties": {
          "response": {
            "type": "string",
            "title": "Response",
            "description": "The chatbot's response"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "Session ID for this conversation"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "Timestamp of response generation"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id",
            "description": "Unique message identifier"
          }
        },
        "type": "object",
        "required": [
          "response",
          "session_id",
          "message_id"
        ],
        "title": "ChatResponse",
        "description": "Response model for chat endpoint.\n\nAttributes:\n    response: The chatbot's response message\n    session_id: The session ID for this conversation\n    timestamp: When the response was generated\n    message_id: Unique identifier for this message exchange"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "APIKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "api_key"
      },
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "token"
          }
        }
      }
    }
  },
...
}